omniagent-ai/requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv

langchain
langchain-community
faiss-cpu
sentence-transformers

transformers
torch

openai-whisper

pymupdf
